+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorInputBoundary.java
@@ -1,4 +1,4 @@
-package generate_maze_use_case;
+package usecases.generateMazeUseCase;
 
 public interface MazeGeneratorInputBoundary {
     MazeGeneratorResponseModel generateMaze(MazeGeneratorRequestModel requestModel);
