+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorInteractor.java
@@ -1,4 +1,4 @@
-package generate_maze_use_case;
+package usecases.generateMazeUseCase;
 
 import entities.MazeFacade;
 import entities.MazeFactory;
