+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorOutputBoundary.java
@@ -1,4 +1,4 @@
-package generate_maze_use_case;
+package usecases.generateMazeUseCase;
 
 public interface MazeGeneratorOutputBoundary {
     MazeGeneratorResponseModel prepareSuccessView(MazeRepresentation mazeRepresentation);
